parte_2/requirements.txt
pandas
sklearn.neighbors 
sklearn.naive_bayes
sklearn.preprocessing 
sklearn.metrics 
sklearn.model_selection 
matplotlib.pyplot 
sklearn.metrics 



import pandas as pd
from sklearn.metrics import accuracy_score
from sklearn.metrics import precision_score
from sklearn.metrics import recall_score
from sklearn.metrics import f1_score
from sklearn.metrics import confusion_matrix
from sklearn.metrics import roc_curve, auc
from sklearn.metrics import roc_auc_score
from sklearn.metrics import mean_squared_error
import matplotlib.pyplot as plt
from sklearn import preprocessing
from sklearn.preprocessing import OrdinalEncoder


from sklearn.ensemble import RandomForestClassifier, RandomForestRegressor
from sklearn.ensemble import ExtraTreesClassifier
import xgboost as xgb
from sklearn.model_selection import train_test_split

import matplotlib.pyplot as plt

